pyproject.toml
[tool.poetry]
name = "image-handler"
version = "0.1.0"
description = "Generate ai images"
authors = [
  "Nathan <probysuper8@gmail.com>",
  "Faisal <faisalfawad2004@gmail.com>"
]
readme = "README.md"
repository = "https://github.com/imaginate-ai/image-handler"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.0"
requests = "^2.0"
marshmallow = "^3.0"
transformers = "^4.0"
aws-lambda-powertools = "^1.0"
torch = { url = "https://download.pytorch.org/whl/cu121/torch-2.3.1%2Bcu121-cp312-cp312-win_amd64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.3.1%2Bcu121-cp312-cp312-win_amd64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu121/torchvision-0.18.1%2Bcu121-cp312-cp312-win_amd64.whl" }
diffusers = "^0.29.2"
pytest = "^8.2.2"
pre-commit = "^3.7.1"
accelerate = "^0.33.0"
ruff = "^0.5.5"
pillow = "^10.4.0"
image-handler-client = {git = "https://github.com/imaginate-ai/image-handler-client.git", rev = "v1.0.2"}
ollama = "^0.3.2"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"

[tool.ruff]
exclude = []
indent-width = 4
target-version = "py312"
src = ["image_handler", "tests"]
line-length = 120

[tool.ruff.lint]
select = [
  "BLE",
  "E",
  "F",
  "I",
  "PL",
  "S",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "D",
  "S",
  "PLR2004",
  "ARG001",
]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["image_handler"]
testpaths = ["tests/unit"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"